src/riot_fetcher/requirements.txt
requests
pymongo
confluent-kafka
python-dotenv
aiokafka
riotwatcher==3.3.1
